gnostr-legit:intermediate:more impl
