+++ asyncgit/src/lib.rs
@@ -47,7 +47,6 @@ mod diff;
 mod error;
 mod fetch_job;
 mod filter_commits;
-pub mod legit;
 mod progress;
 mod pull;
 mod push;
